requirements.txt
langchain
langchain_openai
python-dotenv
openai
faiss-cpu
streamlit
pypdf
docx2txt
openpyxl
fastapi
uvicorn
python-multipart